Move stream and role binding controllers to subpackages
