+++ src/main/java/com/michelin/ns4kafka/controller/rolebindings/RoleBindingController.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.michelin.ns4kafka.controller;
+package com.michelin.ns4kafka.controller.rolebindings;
 
 import static io.micronaut.core.util.StringUtils.EMPTY_STRING;
